refactor: Mover, Mapper를 util class로 변경
