+++ src/main/java/domain/util/ResultMapper.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.util;
 
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -9,7 +9,9 @@ import domain.user.User;
 
 public class ResultMapper {
 
-	public static Map<User, End> map(List<User> users, List<End> ends, List<Integer> indicies) {
+	public static Map<User, End> map(final List<User> users,
+			final List<End> ends,
+			final List<Integer> indicies) {
 		Map<User, End> map = new LinkedHashMap<>();
 		int size = users.size();
 		for (int i = 0; i < size; i++) {
