+++ src/test/java/domain/util/IndexMoverTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.util;
 
 import java.util.List;
 
@@ -10,8 +10,6 @@ import domain.ladder.Ladder;
 import domain.ladder.LadderBuilder;
 import domain.ladder.LadderHeight;
 import domain.ladder.LadderWidth;
-import domain.util.Point;
-import domain.util.PointGenerator;
 
 class IndexMoverTest {
 
@@ -23,15 +21,14 @@ class IndexMoverTest {
 	}
 
 	@Test
-	@DisplayName("0, 1, 2의 index가 1, 0, 2로 변해야 한다.")
+	@DisplayName("0, 1, 2, 3의 index가 1, 0, 3, 2로 변해야 한다.")
 	void indexMoverTest() {
 		int height = 3;
 		int width = 3;
 		PointGenerator generator = new PresentPointGenerator();
 		Ladder ladder = makeLadder(height, width, generator);
 
-		IndexMover indexMover = new IndexMover();
-		List<Integer> indices = indexMover.getMovedIndex(ladder);
+		List<Integer> indices = IndexMover.getMovedIndex(ladder);
 
 		Assertions.assertThat(indices).containsExactly(1, 0, 3, 2);
 	}
@@ -43,9 +40,3 @@ class IndexMoverTest {
 		return builder.build(ladderHeight, ladderWidth, generator);
 	}
 }
-/*
-0 1 2 3
-|o|x|o|
-|o|x|o|
-|o|x|o|
- */
